moved hit files into hit folder
