+++ src/hit/hit_cylinder_caps.c
