+++ src/hit/hit_sphere.c
